O365Connector Actionable Cards Changes
